Create Scores component
