+++ client/src/pages/Game.tsx
@@ -11,6 +11,7 @@ import { RoundResult } from '../types/RoundResult'
 import { CardType } from '../types/CardType'
 import { Card } from '../types/Card'
 import Round from './Round'
+import Scores from './Scores'
 import Lobby from './Lobby'
 import { Settings } from '../types/Settings'
 import Podium from '../components/Podium'
@@ -94,15 +95,20 @@ const Game: FC<GameProps> = (props: GameProps) => {
             />
           )
         case GameStatus.ROUND_IN_PROGRESS:
-        case GameStatus.ROUND_ENDED:
           return (
             <Round
               game={game}
+              cards={cards}
+              hitPile={() => hubConnection.invoke('HitPile')}
+            />
+          )
+        case GameStatus.ROUND_ENDED:
+          return (
+            <Scores
+              game={game}
               currentPlayerId={currentPlayerId}
               isGameHost={isGameHost}
-              cards={cards}
               result={result}
-              hitPile={() => hubConnection.invoke('HitPile')}
               startGame={() => hubConnection.invoke('StartGame')}
             />
           )
