+++ client/src/pages/Round/Round.scss
@@ -5,12 +5,4 @@
     display: block;
     margin: 0 auto $spacing-regular;
   }
-
-  &__Players {
-    margin-top: $spacing-large;
-  }
-
-  &__NewRoundButton {
-    margin-top: $spacing-large;
-  }
 }
