+++ client/src/pages/Scores/index.ts
@@ -1,0 +1,2 @@
+export * from './Scores'
+export { default } from './Scores'
